+++ src/CryptoAPIs/Model/GetTransactionDetailsByTransactionIDRIBSZVin.cs
@@ -47,31 +47,36 @@ namespace CryptoAPIs.Model
         /// <param name="txinwitness">txinwitness (required).</param>
         /// <param name="value">Defines the specific amount. (required).</param>
         /// <param name="vout">It refers to the index of the output address of this transaction. The index starts from 0. (required).</param>
-        public GetTransactionDetailsByTransactionIDRIBSZVin(List<string> addresses = default(List<string>), GetTransactionDetailsByTransactionIDRIBSZScriptSig scriptSig = default(GetTransactionDetailsByTransactionIDRIBSZScriptSig), int sequence = default(int), string txid = default(string), List<string> txinwitness = default(List<string>), string value = default(string), int vout = default(int))
+        public GetTransactionDetailsByTransactionIDRIBSZVin(List<string> addresses = default(List<string>), GetTransactionDetailsByTransactionIDRIBSZScriptSig scriptSig = default(GetTransactionDetailsByTransactionIDRIBSZScriptSig), long sequence = default(long), string txid = default(string), List<string> txinwitness = default(List<string>), string value = default(string), int vout = default(int))
         {
             // to ensure "addresses" is required (not null)
-            if (addresses == null) {
+            if (addresses == null)
+            {
                 throw new ArgumentNullException("addresses is a required property for GetTransactionDetailsByTransactionIDRIBSZVin and cannot be null");
             }
             this.Addresses = addresses;
             // to ensure "scriptSig" is required (not null)
-            if (scriptSig == null) {
+            if (scriptSig == null)
+            {
                 throw new ArgumentNullException("scriptSig is a required property for GetTransactionDetailsByTransactionIDRIBSZVin and cannot be null");
             }
             this.ScriptSig = scriptSig;
             this.Sequence = sequence;
             // to ensure "txid" is required (not null)
-            if (txid == null) {
+            if (txid == null)
+            {
                 throw new ArgumentNullException("txid is a required property for GetTransactionDetailsByTransactionIDRIBSZVin and cannot be null");
             }
             this.Txid = txid;
             // to ensure "txinwitness" is required (not null)
-            if (txinwitness == null) {
+            if (txinwitness == null)
+            {
                 throw new ArgumentNullException("txinwitness is a required property for GetTransactionDetailsByTransactionIDRIBSZVin and cannot be null");
             }
             this.Txinwitness = txinwitness;
             // to ensure "value" is required (not null)
-            if (value == null) {
+            if (value == null)
+            {
                 throw new ArgumentNullException("value is a required property for GetTransactionDetailsByTransactionIDRIBSZVin and cannot be null");
             }
             this.Value = value;
@@ -95,7 +100,7 @@ namespace CryptoAPIs.Model
         /// </summary>
         /// <value>Represents the script sequence number.</value>
         [DataMember(Name = "sequence", IsRequired = true, EmitDefaultValue = false)]
-        public int Sequence { get; set; }
+        public long Sequence { get; set; }
 
         /// <summary>
         /// Represents the reference transaction identifier.
