+++ src/CryptoAPIs/Model/ListAllUnconfirmedTransactionsRIBSD.cs
@@ -45,18 +45,20 @@ namespace CryptoAPIs.Model
         /// <param name="version">Defines the version of the transaction. (required).</param>
         /// <param name="vin">Represents the transaction inputs. (required).</param>
         /// <param name="vout">Represents the transaction outputs. (required).</param>
-        public ListAllUnconfirmedTransactionsRIBSD(int locktime = default(int), int size = default(int), int version = default(int), List<ListAllUnconfirmedTransactionsRIBSDVin> vin = default(List<ListAllUnconfirmedTransactionsRIBSDVin>), List<GetTransactionDetailsByTransactionIDRIBSDVout> vout = default(List<GetTransactionDetailsByTransactionIDRIBSDVout>))
+        public ListAllUnconfirmedTransactionsRIBSD(long locktime = default(long), int size = default(int), int version = default(int), List<ListAllUnconfirmedTransactionsRIBSDVin> vin = default(List<ListAllUnconfirmedTransactionsRIBSDVin>), List<GetTransactionDetailsByTransactionIDRIBSDVout> vout = default(List<GetTransactionDetailsByTransactionIDRIBSDVout>))
         {
             this.Locktime = locktime;
             this.Size = size;
             this._Version = version;
             // to ensure "vin" is required (not null)
-            if (vin == null) {
+            if (vin == null)
+            {
                 throw new ArgumentNullException("vin is a required property for ListAllUnconfirmedTransactionsRIBSD and cannot be null");
             }
             this.Vin = vin;
             // to ensure "vout" is required (not null)
-            if (vout == null) {
+            if (vout == null)
+            {
                 throw new ArgumentNullException("vout is a required property for ListAllUnconfirmedTransactionsRIBSD and cannot be null");
             }
             this.Vout = vout;
@@ -67,7 +69,7 @@ namespace CryptoAPIs.Model
         /// </summary>
         /// <value>Represents the locktime on the transaction on the specific blockchain, i.e. the blockheight at which the transaction is valid.</value>
         [DataMember(Name = "locktime", IsRequired = true, EmitDefaultValue = false)]
-        public int Locktime { get; set; }
+        public long Locktime { get; set; }
 
         /// <summary>
         /// Represents the total size of this transaction.
