+++ src/CryptoAPIs/Model/ListConfirmedTransactionsByAddressRIBSBC.cs
@@ -45,13 +45,14 @@ namespace CryptoAPIs.Model
         /// <param name="version">Represents the transaction&#39;s version number. (required).</param>
         /// <param name="vin">Represents the transaction inputs. (required).</param>
         /// <param name="vout">Represents the transaction outputs..</param>
-        public ListConfirmedTransactionsByAddressRIBSBC(int locktime = default(int), int size = default(int), int version = default(int), List<ListConfirmedTransactionsByAddressRIBSBCVin> vin = default(List<ListConfirmedTransactionsByAddressRIBSBCVin>), List<GetTransactionDetailsByTransactionIDRIBSBCVout> vout = default(List<GetTransactionDetailsByTransactionIDRIBSBCVout>))
+        public ListConfirmedTransactionsByAddressRIBSBC(long locktime = default(long), int size = default(int), int version = default(int), List<ListConfirmedTransactionsByAddressRIBSBCVin> vin = default(List<ListConfirmedTransactionsByAddressRIBSBCVin>), List<GetTransactionDetailsByTransactionIDRIBSBCVout> vout = default(List<GetTransactionDetailsByTransactionIDRIBSBCVout>))
         {
             this.Locktime = locktime;
             this.Size = size;
             this._Version = version;
             // to ensure "vin" is required (not null)
-            if (vin == null) {
+            if (vin == null)
+            {
                 throw new ArgumentNullException("vin is a required property for ListConfirmedTransactionsByAddressRIBSBC and cannot be null");
             }
             this.Vin = vin;
@@ -63,7 +64,7 @@ namespace CryptoAPIs.Model
         /// </summary>
         /// <value>Represents the locktime on the transaction on the specific blockchain, i.e. the blockheight at which the transaction is valid.</value>
         [DataMember(Name = "locktime", IsRequired = true, EmitDefaultValue = false)]
-        public int Locktime { get; set; }
+        public long Locktime { get; set; }
 
         /// <summary>
         /// Represents the total size of this transaction.
